添加防止重复提交，异常处理功能

+++ src/main/java/com/netease/shopping/serviceImpl/CommodityServiceImpl.java
@@ -2,12 +2,15 @@ package com.netease.shopping.serviceImpl;
 
 import com.netease.shopping.dao.mapper.CommodityMapper;
 import com.netease.shopping.entity.Commodity;
+import com.netease.shopping.entity.ShoppingCart;
 import com.netease.shopping.service.CommodityService;
 
+import com.netease.shopping.service.ShoppingCartService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -23,20 +26,29 @@ public class CommodityServiceImpl implements CommodityService{
     @Autowired
     private CommodityMapper commodityMapper;
 
+    @Autowired
+    private ShoppingCartService shoppingCartService;
+
     private static final Logger log = LoggerFactory.getLogger(CommodityServiceImpl.class);
 
     public Integer insertCommodity(Commodity commodity) {
         return  commodityMapper.insertCommodity(commodity);
     }
 
+    @Transactional
     public Integer deleteCommodityById(int commodityID) {
         if(commodityID<0){
             return ER5001;
         }
-        if(commodityMapper.deleteCommodityById(commodityID)!=1){
-            log.debug("delete fail commodityID is"+commodityID);
-            return ER6001;
+        List<ShoppingCart> shoppingCartList=shoppingCartService.findShoppingCartByCommodityID(commodityID);
+        if(shoppingCartList!=null&&shoppingCartList.size()>0){
+            shoppingCartService.deleteShoppingCartByCommodityID(commodityID);
+            if(commodityMapper.deleteCommodityById(commodityID)!=1){
+                log.debug("delete fail commodityID is"+commodityID);
+                return ER6001;
+            }
         }
+
         return S1000;
     }
